refactor: share EditNominationForm component
